ml_model/requirements.txt
pandas
flask
numpy
mlflow
scikit-learn==1.6.1
azure-storage-blob